Extract rigid object interface test utilities
Move the rigid object and rigid object collection interface test
factories into shared underscore-prefixed utility modules so other
interface test files can reuse them. Pure code movement with no
behavior change; part 1 of the articulation reordering series.
